move plancard into components
